use a placeholder value for :github-client-id and :github-client-secret in handler.clj
The upgrade from `ring-oauth2` from 0.1.4 to 0.1.5 adds a change here:
https://github.com/weavejester/ring-oauth2/blob/5bf6df4e9afa1f1362581417c23b03867891e5f3/src/ring/middleware/oauth2.clj#L120
which validates the oauth2 header against a clojure spec that complains
whenever a defined client id or client secret is nil.

DROID already complains when the github client secret and client id are
not defined, but this error only goes to the log and does not crash the
application. It is important that it not crash because at compile time
the github client secret and id are not known.

To workaround the issue we store empty strings in the header instead of
nil.